Delete excess file. complete User constructor
